merge

+++ src/main/java/com/dbc/vemserback/ecommerce/dto/PurchaseList/PurchaseDTO.java
@@ -1,9 +1,5 @@
 package com.dbc.vemserback.ecommerce.dto.PurchaseList;
 
-import java.math.BigDecimal;
-
-import org.springframework.web.multipart.MultipartFile;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,9 +10,7 @@ import lombok.NoArgsConstructor;
 @Data
 @Builder
 public class PurchaseDTO {
-	
 	private String name;
 	private String description;
-	private BigDecimal price;
-	
+	private Double price;
 }
